+++ src/Molecules/Select/ValueDisplayMultiSelect/ValueDisplayMultiSelect.types.tsx
@@ -1,0 +1,5 @@
+export type Props = {
+  label: string;
+  placeholder: string;
+  selectedCount: number;
+};